+++ R/module-chartControls.R
@@ -83,8 +83,13 @@ chartControlsUI <- function(id) {
 #' @param data_table \code{reactive} function returning data used in plot.
 #' @param data_name \code{reactive} function returning data name.
 #' @param ggplot_rv \code{reactiveValues} with ggplot object (for export).
+#' @param aesthetics \code{reactive} function returning aesthetic names used.
 #' @param use_facet \code{reactive} function returning
 #'  \code{TRUE} / \code{FALSE} if plot use facets.
+#' @param use_transX \code{reactive} function returning \code{TRUE} / \code{FALSE}
+#'  to use transformation on x-axis.
+#' @param use_transY \code{reactive} function returning \code{TRUE} / \code{FALSE}
+#'  to use transformation on y-axis.
 #'
 #' @return A reactiveValues with all input's values
 #' @noRd
@@ -98,6 +103,7 @@ chartControlsUI <- function(id) {
 chartControlsServer <- function(input, output, session, 
                                 type, data_table, data_name,
                                 ggplot_rv, 
+                                aesthetics = reactive(NULL),
                                 use_facet = shiny::reactive(FALSE), 
                                 use_transX = shiny::reactive(FALSE), 
                                 use_transY = shiny::reactive(FALSE)) {
@@ -182,6 +188,25 @@ chartControlsServer <- function(input, output, session,
   
   
   # Controls ----
+  
+  observeEvent(aesthetics(), {
+    aesthetics <- aesthetics()
+    if ("fill" %in% aesthetics) {
+      toggleDisplay(id = ns("controls-labs-fill"), display = "block")
+    } else {
+      toggleDisplay(id = ns("controls-labs-fill"), display = "none")
+    }
+    if ("color" %in% aesthetics) {
+      toggleDisplay(id = ns("controls-labs-color"), display = "block")
+    } else {
+      toggleDisplay(id = ns("controls-labs-color"), display = "none")
+    }
+    if ("size" %in% aesthetics) {
+      toggleDisplay(id = ns("controls-labs-size"), display = "block")
+    } else {
+      toggleDisplay(id = ns("controls-labs-size"), display = "none")
+    }
+  })
   
   observeEvent(use_facet(), {
     if (isTRUE(use_facet())) {
@@ -279,12 +304,19 @@ chartControlsServer <- function(input, output, session,
   
   # labs input
   observe({
+    asth <- aesthetics()
+    labs_fill <- ifelse("fill" %in% asth, input$labs_fill, "")
+    labs_color <- ifelse("color" %in% asth, input$labs_color, "")
+    labs_size <- ifelse("size" %in% asth, input$labs_size, "")
     outin$labs <- list(
       x = input$labs_x %empty% NULL,
       y = input$labs_y %empty% NULL,
       title = input$labs_title %empty% NULL,
       subtitle = input$labs_subtitle %empty% NULL,
-      caption = input$labs_caption %empty% NULL
+      caption = input$labs_caption %empty% NULL,
+      fill = labs_fill %empty% NULL,
+      color = labs_color %empty% NULL,
+      size = labs_size %empty% NULL
     )
   })
   
@@ -366,7 +398,19 @@ controls_labs <- function(ns) {
     textInput(inputId = ns("labs_subtitle"), placeholder = "Subtitle", label = NULL),
     textInput(inputId = ns("labs_caption"), placeholder = "Caption", label = NULL),
     textInput(inputId = ns("labs_x"), placeholder = "X label", label = NULL),
-    textInput(inputId = ns("labs_y"), placeholder = "Y label", label = NULL)
+    textInput(inputId = ns("labs_y"), placeholder = "Y label", label = NULL),
+    tags$div(
+      id = ns("controls-labs-fill"), style = "display: none;",
+      textInput(inputId = ns("labs_fill"), placeholder = "Fill label", label = NULL)
+    ),
+    tags$div(
+      id = ns("controls-labs-color"), style = "display: none;",
+      textInput(inputId = ns("labs_color"), placeholder = "Color label", label = NULL)
+    ),
+    tags$div(
+      id = ns("controls-labs-size"), style = "display: none;",
+      textInput(inputId = ns("labs_size"), placeholder = "Size label", label = NULL)
+    )
   )
 }
 
@@ -389,13 +433,18 @@ controls_labs <- function(ns) {
 #'  theme_excel theme_few theme_fivethirtyeight theme_foundation theme_gdocs theme_hc 
 #'  theme_igray theme_map theme_pander theme_par theme_solarized theme_solarized_2 
 #'  theme_solid theme_stata theme_tufte theme_wsj
+#' @importFrom hrbrthemes theme_ft_rc theme_ipsum theme_ipsum_ps theme_ipsum_rc
+#'  theme_ipsum_tw theme_modern_rc
 controls_appearance <- function(ns) {
 
   themes <- list(
     ggplot2 = list(
       "bw", "classic", "dark", "gray",
-      "grey", "light", "linedraw", "minimal",
+      "light", "linedraw", "minimal",
       "void"
+    ),
+    hrbrthemes = c(
+      "ft_rc", "ipsum", "ipsum_ps", "ipsum_rc", "ipsum_tw", "modern_rc"
     ),
     ggthemes = list(
       "base", "calc", "economist", "economist_white",
@@ -651,6 +700,7 @@ controls_code <- function(ns) {
 #'
 #' @importFrom RColorBrewer brewer.pal brewer.pal.info
 #' @importFrom scales hue_pal viridis_pal
+#' @importFrom hrbrthemes ipsum_pal ft_pal
 colors_palettes <- function() {
   ### colors
   # For spectrum pre-defined colors
@@ -660,6 +710,9 @@ colors_palettes <- function() {
     "inferno" = col2Hex(viridis_pal(option = "inferno")(10)),
     "plasma" = col2Hex(viridis_pal(option = "plasma")(10)),
     "cividis" = col2Hex(viridis_pal(option = "cividis")(10))
+    ,
+    "ipsum" = ipsum_pal()(9),
+    "ft" = ft_pal()(9)
     ,
     "Blues" = get_brewer_pal(name = "Blues"),
     "Greens" = get_brewer_pal(name = "Greens"),
@@ -700,16 +753,21 @@ colors_palettes <- function() {
       "plasma" = col2Hex(viridis_pal(option = "plasma")(10)),
       "cividis" = col2Hex(viridis_pal(option = "cividis")(10))
     ),
+    list(
+      "ipsum" = ipsum_pal()(9),
+      "ft" = ft_pal()(9)
+    ),
     background_pals
   )
   background_pals <- unlist(lapply(X = background_pals, FUN = linear_gradient))
   colortext_pals <- rep(c("white", "black", "black"), times = sapply(colors_pal, length))
-  colortext_pals <- c("white", rep("white", 5), colortext_pals) # ggplot2 + viridis
+  colortext_pals <- c("white", rep("white", 5), rep("white", 2), colortext_pals) # ggplot2 + viridis + hrbrthemes
 
   # add ggplot2 hue & viridis
   colors_pal <- c(
     list("Default" = list("ggplot2")),
     list("Viridis" = list("viridis", "magma", "inferno", "plasma", "cividis")),
+    list("hrbrthemes" = list("ipsum", "ft")),
     colors_pal
   )